update application page
