+++ frontend/app/applications/[id]/ApplicationHeader.js
@@ -1,23 +1,45 @@
 'use client'
-import { Typography, Stack, Link, Box } from '@mui/material'
+import { Typography, Stack, Link, Box, IconButton, Paper } from '@mui/material'
 import { useApplicationContext } from '@/components/Context/ApplicationContext'
+import EditIcon from '@mui/icons-material/Edit'
 
 export default function ApplicationHeader() {
   const { application } = useApplicationContext()
 
   return (
     <Box>
-      <Typography
-        component="h1"
+      <Stack
         sx={{
-          fontSize: { xs: '1.5rem', sm: '1.75rem' },
-          pb: { xs: 0.5, sm: 2 },
-          paddingX: 2,
-          fontWeight: 600,
+          justifyContent: 'space-between',
+          alignItems: 'start',
+          flexDirection: { xs: 'column', sm: 'row' },
         }}
       >
-        {application.job_title}
-      </Typography>
+        <Typography
+          component="h1"
+          sx={{
+            fontSize: { xs: '1.5rem', sm: '1.75rem' },
+            pb: { xs: 0.5, sm: 2 },
+            paddingX: 2,
+            fontWeight: 600,
+          }}
+        >
+          {application.job_title}
+        </Typography>
+        <Paper sx={{ marginLeft: { xs: 2 }, mb: { xs: 1 } }}>
+          <Typography
+            component="h2"
+            sx={{
+              fontSize: { xs: '1rem', sm: '1.2rem' },
+              p: { xs: 0.5, sm: 1 },
+              paddingX: 2,
+              fontWeight: 600,
+            }}
+          >
+            {application.status.toLocaleUpperCase()}
+          </Typography>
+        </Paper>
+      </Stack>
 
       <Stack
         sx={{
@@ -35,7 +57,7 @@ export default function ApplicationHeader() {
             }}
           >
             <Typography
-              component="h2"
+              component="h3"
               sx={{ fontSize: { xs: '1rem', sm: '1.2rem' } }}
             >
               {application.company_name}
@@ -52,6 +74,9 @@ export default function ApplicationHeader() {
                 location
               </Typography>
             )}
+            <IconButton>
+              <EditIcon sx={{ color: 'secondary.main', fontSize: 20 }} />
+            </IconButton>
           </Stack>
 
           {application.company_website ? (
